backend/pyproject.toml
[project]
name = "curecompass-api"
version = "0.1.0"
description = "CureCompass backend API and ingestion services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "email-validator>=2.2.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.32",
  "alembic>=1.13.2",
  "psycopg[binary]>=3.2.1",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt>=4.1.0",
  "redis>=5.0.8",
  "celery>=5.4.0",
  "httpx>=0.27.0",
  "tenacity>=8.5.0",
  "openai>=1.45.0",
  "jiter>=0.8.0",
  "pgvector>=0.3.2",
  "orjson>=3.10.7",
  "python-multipart>=0.0.9",
  "markdown>=3.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.4",
  "mypy>=1.11.2",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
